oscar_assignment/requirements.txt
Babel==2.6.0
Django==2.1.8
django-appconf==1.0.3
django-compressor==2.2
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==1.6.7
django-oscar-api==1.5.4
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.5
gunicorn==19.9.0
mock==2.0.0
pbr==5.1.3
phonenumbers==8.10.10
Pillow==6.0.0
pip==19.1
psycopg2==2.8.2
purl==1.5
python-dateutil==2.8.0
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.0.12
setuptools==41.0.1
six==1.12.0
sorl-thumbnail==12.4.1
sqlparse==0.3.0
text-unidecode==1.2
Unidecode==1.0.23
wheel==0.33.1
